warning: implicit conversion to float (#1824)

+++ src/sound/SoundDevice.cc
@@ -103,7 +103,7 @@ void SoundDevice::registerSound(const DeviceConfig& config)
 
 		auto channels = StringOp::parseRange(channel->getValue(), 1, numChannels);
 		channels.foreachSetBit([&](size_t c) {
-			setBalance(unsigned(c - 1), *balance * (1.0f / 100.0f));
+			setBalance(unsigned(c - 1), narrow_cast<float>(*balance) * (1.0f / 100.0f));
 		});
 	}
